+++ CHANGELOG.md
@@ -9,3 +9,11 @@
   * react-apollo
   * redux
   * graphql
+  
+## [0.0.2] - 2018-02-02
+
+### Added
+
+- Added Cognito SDKs.
+- Added Login and Signup page.
+- Added Forgot password and Reset password page.